refactor(query): split query and query utils
