+++ src/query/datasetquery.ts
@@ -1,222 +1,12 @@
-import { literal, namedNode, variable } from "@rdfjs/data-model";
+import { namedNode, variable } from "@rdfjs/data-model";
 import clone from "clone";
 import { Generator as SparqlGenerator } from "sparqljs";
-import Component from "./components/index";
-import DataSet from "./dataset";
-import Binding from "./expressions/binding";
-import Operator from "./expressions/operator";
-import { ArrayExpr, IExpr, into, isTerm, TermExpr } from "./expressions/utils";
-import SparqlFetcher from "./sparqlfetcher";
-import { BgpPattern, Expression, FilterPattern, OperationExpression, Ordering, SelectQuery, Tuple } from "./sparqljs.d";
-
-type PredicateFunction = (data: Selects) => Component;
-type Selects = Record<string, Component>;
-
-interface IState {
-  selects: Selects;
-  filters: IExpr[];
-  groupBys: Array<PredicateFunction | string>;
-  havings: PredicateFunction[];
-  offset: number;
-  limit: number;
-  order: Component[];
-}
-
-export interface IQueryOpts {
-  languages?: string[];
-}
-
-const baseState: IState = {
-  selects: {},
-  filters: [],
-  groupBys: [],
-  havings: [],
-  offset: 0,
-  limit: 10,
-  order: [],
-};
-
-/**
- * Convert [[Operator]] arguments into SPARQL.js `Expression`s
- *
- * @param {Operator} operator
- */
-function operatorArgsToExpressions(args: IExpr[]): Expression[] {
-  const expressions = args.map((arg: IExpr): Expression => {
-    if (isTerm(arg)) {
-      return arg;
-    }
-    if (arg instanceof Binding) {
-      return variable(arg.name);
-    }
-    if (arg instanceof Operator) {
-      return createOperationExpression(arg);
-    }
-    if (arg instanceof TermExpr) {
-      return arg.term;
-    }
-    if (arg instanceof ArrayExpr) {
-      const tuple: Tuple = operatorArgsToExpressions(Array.from(arg.xs).map(into));
-      return tuple;
-    }
-  }).filter((x) => {
-    const transformed = Boolean(x);
-    if (!transformed) {
-      throw new Error("Unrecognized filter argument type");
-    }
-    return transformed;
-  });
-  return expressions;
-}
-
-function createOperationExpression(operator: Operator): OperationExpression {
-  const operationExpression: OperationExpression = {
-    type: "operation",
-    operator: operator.operator,
-    args: operatorArgsToExpressions(operator.args),
-  };
-  return operationExpression;
-}
-
-/**
- * When `.filter` is called several times, apply a logical AND between all filters.
- */
-function combineFilters(operations: OperationExpression[]): FilterPattern {
-  let combined;
-  if (operations.length > 1) {
-    combined = operations
-      .reduce((acc, op) => {
-        acc.args.push(op);
-        return acc;
-      }, {
-        operator: "&&",
-        type: "operation",
-        args: [],
-      });
-  } else {
-    combined = operations[0];
-  }
-  return {
-    type: "filter",
-    expression: combined,
-  };
-}
-
-function langPrepare(binding, labelBinding, labelLangBinding, langs: string[]) {
-  // fetch labels when they exist, handling languages
-  const langMatch = (lang: string): OperationExpression => {
-    return {
-      type: "operation",
-      operator: "langmatches",
-      args: [
-        {
-          type: "operation",
-          operator: "lang",
-          args: [ labelLangBinding ],
-        },
-        literal(lang),
-      ],
-    };
-  };
-
-  const langExactMatch = (lang = ""): OperationExpression => {
-    return {
-      type: "operation",
-      operator: "=",
-      args: [
-        {
-          type: "operation",
-          operator: "lang",
-          args: [ labelLangBinding ],
-        },
-        literal(lang),
-      ],
-    };
-  };
-
-  function langHandler() {
-    let languages = langs.filter(Boolean);
-    if (!languages.length) {
-      // no lang specified, default to 'empty' lang
-      return {
-        type: "filter",
-        expression: langExactMatch(""),
-      };
-    }
-    // at least one non-empty lang, add the 'empty' lang as fallback
-    languages = langs.concat("");
-    // we now have at least two langs to work with
-    const lang1 = languages.shift();
-    const lang2 = languages.shift();
-    let expression: OperationExpression = {
-      type: "operation",
-      operator: "||",
-      args: [
-        lang1 === "" ? langExactMatch(lang1) : langMatch(lang1),
-        lang2 === "" ? langExactMatch(lang2) : langMatch(lang2),
-      ],
-    };
-
-    let extraLang = languages.shift();
-    while (typeof extraLang !== "undefined") {
-      expression = {
-        type: "operation",
-        operator: "||",
-        args: [
-          expression,
-          extraLang === "" ? langExactMatch("") : langMatch(extraLang),
-        ],
-      };
-      extraLang = languages.shift();
-    }
-    return {
-      type: "filter",
-      expression,
-    };
-  }
-
-  const findLabel = {
-    type: "optional",
-    patterns: [
-      {
-        type: "bgp",
-        triples: [
-          {
-            subject: binding,
-            predicate: {
-              type: "path",
-              pathType: "|",
-              items: [
-                namedNode("http://www.w3.org/2000/01/rdf-schema#label"),
-                namedNode("http://www.w3.org/2004/02/skos/core#prefLabel"),
-              ],
-            },
-            object: labelLangBinding,
-          },
-        ],
-      },
-      langHandler(),
-    ],
-  };
-
-  const coalesceLabel = {
-    type: "bind",
-    variable: labelBinding,
-    expression: {
-      type: "operation",
-      operator: "coalesce",
-      args: [
-        labelLangBinding,
-        binding,
-      ],
-    },
-  };
-
-  return {
-    findLabel,
-    coalesceLabel,
-  };
-}
+import Component from "../components/index";
+import DataSet from "../dataset";
+import { IExpr } from "../expressions/utils";
+import SparqlFetcher from "../sparqlfetcher";
+import { BgpPattern, FilterPattern, Ordering, SelectQuery } from "../sparqljs";
+import { baseState, combineFilters, createOperationExpression, IQueryOpts, IState, langPrepare, PredicateFunction } from "./utils";
 
 /**
  * A query to a [[DataSet]].
@@ -372,7 +162,6 @@ class DataSetQuery {
   }
 
   public applyFilters(): FilterPattern {
-    const selects = this.state.selects;
     const filters = this.state.filters
       .map((op) => op.resolve(this.iriToBinding))
       .map(createOperationExpression);
